Added Token Display and ReRoll (incomplete)
Added equipped tokens to display. Tokens will be displayed in the inventory screen, as well as the forge. This displays the slot equipped the token type, and the power level of the token.
Added token reroll at the forge. When at the forge, click "Buy Token" to randomly reroll all 3 spots. This does not cost any money.